Improved submerge event on Ouro Added missing timers for initial sweep and blast to Ouro on engage and initial timer for sweep and blast after a submerge.
NOTE: Ouro's Submerge does not have a timer. Check for yourselves on Warcraftlogs. I looked through 30 pulls. not a single one of them had submerge at same time. In fact the variance was so massive I can only conclude it truly is random. Some pulls with submerge as early as 70sec from engage and some with it as late as 3 minutes into fight. It's also worth noting that in most kills he didn't even submerge at all (likely because if the submerge window is 1-3 minutes, and kill is 1minute 30 seconds, there is a good change he died before submerge)